pyproject.toml
[tool.poetry]
name = "netsuite-python"
version = "0.1.0"
description = "Interface for working with the Netsuite Rest Web Services API"
authors = ["Joseph DeBartola"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
PyJWT = "^2.4.0"
requests = "^2.28.0"
cryptography = "^37.0.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
